requirements.txt
absl-py==0.10.0
astor==0.8.1
blinker==1.4
dicom2nifti==2.2.10
gast==0.2.2
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
importlib-metadata==1.7.0
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
nibabel==3.1.1
nilearn==0.6.2
numpy==1.19.1
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.2
protobuf==3.13.0
pydicom==2.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
tensorboard==1.15.0
tensorflow-cpu==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl==2.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
